Rename package to alertaclient
