+++ .travis.yml
@@ -12,5 +12,5 @@ install:
   - "pip install -r requirements.txt"
 
 # command to run tests
-script: nosetests tests/test_alert.py tests/test_alertdoc.py
+script: nosetests tests/*
