--- alerta/__init__.py
@@ -1,2 +1,0 @@
-import pkg_resources
-pkg_resources.declare_namespace(__name__)
