+++ setup.py
@@ -10,15 +10,14 @@ with open('README.md') as f:
 
 setuptools.setup(
     name="alerta",
-    namespace_packages=['alerta'],
     version=version,
-    description="Alerta unified command-line tool",
+    description="Alerta unified command-line tool and SDK",
     long_description=readme,
     license="MIT",
     author="Nick Satterly",
     author_email="nick.satterly@theguardian.com",
-    url="http://github.com/alerta/python-alerta-client",
-    packages=['alerta'],
+    url="http://github.com/alerta/python-alerta",
+    packages=['alertaclient'],
     install_requires=[
         'argparse',
         'requests',
@@ -29,11 +28,12 @@ setuptools.setup(
             'alerta = alerta.shell:main'
         ]
     },
-    keywords="alerta client unified command line tool",
+    keywords="alerta client unified command line tool sdk",
     classifiers=[
         'Development Status :: 5 - Production/Stable',
         'License :: OSI Approved :: MIT License',
         'Intended Audience :: System Administrators',
+        'Programming Language :: Python :: 3.5',
         'Programming Language :: Python :: 3.4',
         'Programming Language :: Python :: 2.7',
         'Topic :: System :: Monitoring',
